commisiones edit

+++ TurismoGlobalArandas/Models/Commissions.cs
@@ -9,9 +9,9 @@ namespace TurismoGlobalArandas.Models
         public int CommissionId { get; set; }
         public int ProviderId { get; set; }
         public string? Color { get; set; }
-        public double? CommissionAgency { get; set; }
-        public double? CommissionClient { get; set; }
-        public double? CommissionEmployee { get; set; }
+        public decimal? CommissionAgency { get; set; }
+        public decimal? CommissionClient { get; set; }
+        public decimal? CommissionEmployee { get; set; }
         public string? Description { get; set; }
         public bool IsDeleted { get; set; }
